move shared function and class code to parent module
