+++ crates/ruff_linter/src/rules/pyupgrade/rules/pep695/use_pep695_type_parameter.rs
@@ -1,15 +1,13 @@
-use itertools::Itertools;
-
 use ruff_diagnostics::{Applicability, Diagnostic, Edit, Fix, FixAvailability, Violation};
 use ruff_macros::{derive_message_formats, ViolationMetadata};
 use ruff_python_ast::{visitor::Visitor, Expr, ExprSubscript};
-use ruff_python_ast::{Stmt, StmtClassDef, StmtFunctionDef};
+use ruff_python_ast::{StmtClassDef, StmtFunctionDef};
 use ruff_text_size::{Ranged, TextRange};
 
 use crate::checkers::ast::Checker;
 use crate::settings::types::PythonVersion;
 
-use super::{DisplayTypeVars, TypeVar, TypeVarReferenceVisitor};
+use super::{check_type_vars, in_nested_context, DisplayTypeVars, TypeVarReferenceVisitor};
 
 /// ## What it does
 ///
@@ -93,15 +91,6 @@ impl Violation for NonPEP695TypeParameter {
     }
 }
 
-/// Check if the current statement is nested within another [`StmtClassDef`] or [`StmtFunctionDef`].
-fn in_nested_context(checker: &Checker) -> bool {
-    checker
-        .semantic()
-        .current_statements()
-        .skip(1) // skip the immediate parent, we only call this within a class or function
-        .any(|stmt| matches!(stmt, Stmt::ClassDef(_) | Stmt::FunctionDef(_)))
-}
-
 /// UP046
 pub(crate) fn non_pep695_generic_class(checker: &mut Checker, class_def: &StmtClassDef) {
     // PEP-695 syntax is only available on Python 3.12+
@@ -179,27 +168,6 @@ pub(crate) fn non_pep695_generic_class(checker: &mut Checker, class_def: &StmtCl
             Applicability::Safe,
         )),
     );
-}
-
-/// Deduplicate `vars`, returning `None` if `vars` is empty or any duplicates are found.
-fn check_type_vars(vars: Vec<TypeVar<'_>>) -> Option<Vec<TypeVar<'_>>> {
-    if vars.is_empty() {
-        return None;
-    }
-
-    // Type variables must be unique; filter while preserving order.
-    let nvars = vars.len();
-    let type_vars = vars
-        .into_iter()
-        .unique_by(|TypeVar { name, .. }| name.id.as_str())
-        .collect::<Vec<_>>();
-
-    // non-unique type variables are runtime errors, so just bail out here
-    if type_vars.len() < nvars {
-        return None;
-    }
-
-    Some(type_vars)
 }
 
 /// UP046
